Add IdentitySeeder

+++ MNF-PORTAL-Service/Services/RoleService.cs
@@ -78,6 +78,10 @@ namespace MNF_PORTAL_Service.Services
             {
                 throw new ArgumentException($"The role '{roleName}' does not exist.");
             }
+            if (roleName == "SuperAdmin")
+            {
+                throw new InvalidOperationException(message: "SuperAdmin role cannot be deleted.");
+            }
             var result = await unitOfWork.RoleRepository.RemoveRoleAsync(roleName);
             if (!result)
             {
@@ -95,6 +99,10 @@ namespace MNF_PORTAL_Service.Services
             if (!await unitOfWork.RoleRepository.RoleIsExistAsync(model.OldRole))
             {
                 throw new ArgumentException($"The role '{model.OldRole}' does not exist.");
+            }
+            if (model.OldRole == "SuperAdmin")
+            {
+                throw new InvalidOperationException(message: "SuperAdmin role cannot be updated.");
             }
             var result = await unitOfWork.RoleRepository.UpdateRoleAsync(model.OldRole, model.NewRole);
             if (!result)
